chore: Install standard-version module (#274)
standard-version is used to help with automatic CHANGELOG
generation and publishing to npm. It replaces the scripts
that were used to help manually generate the CHANGELOG.

The available commands are as follows:

- `version`: Considered the "default" command to create a
new release, this command will automatically decide which
version number to bump (patch, minor, major) and will
require a manual push and publish
- `version:patch`: This command bypasses the automated
version bump and allows you to force a patch release
- `version:minor`: This command bypasses the automated
version bump and allows you to force a minor release
- `version:major`: This command bypasses the automated
version bump and allows you to force a major release
- `push`: This command is an alias to push and publish if you
use one of the `version` commands
- `deploy: If you want to release, push and publish all at
once, you can run this command. It works in the same way as
`version` in that it will automatically bump the version but
it will also run the `push` command afterwards
`deploy:patch`: Automatically push and publish after
creating a patch release
`deploy:minor`: Automatically push and publish after
creating a minor release
`deploy:major`: Automatically push and publish after
creating a major release

Closes #173