update: Update styles for Modal component (#280)
* Change modalPadding value from 56 to 24

* Check for title to render Header

* Conditionally add bottom border to header

There shouldn't be a border under the header if there is no
title passed in. For example, the log in modal doesn't have
a title or a border. Also, by using a max-width media query,
this prevents from having to reset `borderBottom` to `0` on
screens above 768px.

* Remove uppercase text transform on header

This is redundant because the Heading component has its caps
prop set to true and this makes the text uppercase.

* Remove left text-align from header

This doesn't need to be reset because there is no text in the
header on desktop. Otherwise, this should be converted to use
a max-width media query to avoid having to reset text-align.

* Replace padding with height on header

Padding has been placed on the buttons within the header to
increase their tap area.

The height values are a calculation of the desired button
height (24px) plus the vertical padding values (16px mobile,
24px desktop).

* Put padding on buttons, zero out position

Because there is padding on the buttons, their positions
(top, left, right) can be 0.

* Change padding on contentContainer

Done because the design spec has 40px of padding around the
content

* Make mobileTitle inline-block and add top padding

inline-block is used instead of block so that the title only
takes up as much width as it needs based on the amount of
text. Otherwise, the title spans the entire width of the
available space and it doesn't really need to.

Top padding is used to push the text down and center it
vertically.

* Add newline before return block