Update z-index, fix typo in Modal (#294)
* Update z-index for Modal

Change "overlay" to use modal z-index and remove z-index completely
from the modal content. This is because the overlay wraps the modal
content so its z-index should be "modal" instead of "overlay" because
otherwise, the wrapper would have a lower z-index than the content.

* Remove transform transition from Modal overlay

The overlay doesn't transform so the transition isn't needed.

* Fix typo

* Remove important declaration

With the typo in “translate” fixed, the transform property is now
applied and `!important` pushes the modal content down the page too
far. Removing `!important` fixes this.